init utmParamsCapture
