+++ packages/analytics/package.json
@@ -4,7 +4,8 @@
   "version": "0.3.1",
   "author": "HashiCorp",
   "dependencies": {
-    "fathom-client": "^3.2.0"
+    "fathom-client": "^3.2.0",
+    "js-cookie": "^2.2.1"
   },
   "peerDependencies": {
     "react": ">= 16.x"
